Replace PP with CRuby version
